+++ html-repository/source/uim/platform/html_repository/application/use_cases/manage_app_routes.d
@@ -76,7 +76,7 @@ class ManageAppRoutesUseCase : UIMUseCase {
         repo.remove(id);
     }
 
-    long countByApp(HtmlAppId appId) {
+    size_t countByApp(HtmlAppId appId) {
         return repo.countByApp(appId);
     }
