+++ job-scheduling/source/uim/platform/job_scheduling/domain/ports/repositories/run_logs.d
@@ -15,5 +15,5 @@ interface RunLogRepository {
     RunLog[] findByStatus(RunStatus status, JobId jobtenantId, id tenantId);
     void save(RunLog r);
     void update(RunLog r);
-    long countBySchedule(ScheduleId scheduleId);
+    size_t countBySchedule(ScheduleId scheduleId);
 }
