+++ mobile/source/uim/platform/mobile/infrastructure/persistence/memory/device_registrations.d
@@ -53,11 +53,11 @@ class MemoryDeviceRegistrationRepository : DeviceRegistrationRepository {
     store.remove(id);
   }
 
-  long countByApp(MobileAppId appId) {
+  size_t countByApp(MobileAppId appId) {
     return cast(long) store.values.filter!(r => r.appId == appId).array.length;
   }
 
-  long countByTenant(TenantId tenantId) {
+  size_t countByTenant(TenantId tenantId) {
     return cast(long) store.values.filter!(r => r.tenantId == tenantId).array.length;
   }
 }
