+++ mobile/source/uim/platform/mobile/infrastructure/persistence/memory/feature_restrictions.d
@@ -53,7 +53,7 @@ class MemoryFeatureRestrictionRepository : FeatureRestrictionRepository {
     store.remove(id);
   }
 
-  long countByApp(MobileAppId appId) {
+  size_t countByApp(MobileAppId appId) {
     return cast(long) store.values.filter!(r => r.appId == appId).array.length;
   }
 }
