+++ mobile/source/uim/platform/mobile/infrastructure/persistence/memory/mobile_apps.d
@@ -49,7 +49,7 @@ class MemoryMobileAppRepository : MobileAppRepository {
     store.remove(id);
   }
 
-  long countByTenant(TenantId tenantId) {
+  size_t countByTenant(TenantId tenantId) {
     return cast(long) store.values.filter!(a => a.tenantId == tenantId).array.length;
   }
 }
